spotify_outreach/requirements.txt
anthropic>=0.40.0
spotipy>=2.23.0
python-dotenv>=1.0.0
schedule>=1.2.0
